+++ src/gui.h
@@ -2,5 +2,5 @@
 
 namespace gui {
 	extern bool setup();
-	extern void update(float* curV, float tarV);
+	extern void update(int* curV, int tarV);
 }
